third_party/WebKit/LayoutTests/fast/mediastream/MediaStreamConstructor.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("Tests MediaStream constructor.");

var undefined;
var stream;
var newStream;
var nAudio;
var nVideo;
var idArray = new Array();

function checkIdAttribute(id) {
    if (id.length != 36)
        return false;
    for (var i = 0; i < id.length; i++) {
        charCode = id.charCodeAt(i);
        if (charCode == 0x21
            || (charCode >= 0x23 && charCode <= 0x27)
            || (charCode >= 0x2A && charCode <= 0x2B)
            || (charCode >= 0x2D && charCode <= 0x2E)
            || (charCode >= 0x30 && charCode <= 0x39)
            || (charCode >= 0x41 && charCode <= 0x5A)
            || (charCode >= 0x5E && charCode <= 0x7E))
            continue;
        return false;
    }
    for (var i = 0; i < idArray.length; i++) {
        if (id === idArray[i])
            return false;
    }
    idArray.push(id);

    return true;
}

function error() {
    testFailed('Stream generation failed.');
    finishJSTest();
}

function testInterfaces() {
  debug('Interface tests.');
  shouldBe('typeof MediaStream', "'function'");
  shouldBe('MediaStream.length', '0');
  shouldBe('MediaStream.name', "'MediaStream'");
  shouldBe('Object.getPrototypeOf(MediaStream.prototype)', 'EventTarget.prototype');
  // webkitMediaStream is an alias of MediaStream
  shouldBe('window.MediaStream', 'window.webkitMediaStream');
}

function getUserMedia(dictionary, callback) {
    try {
        navigator.webkitGetUserMedia(dictionary, callback, error);
    } catch (e) {
        testFailed('webkitGetUserMedia threw exception :' + e);
        finishJSTest();
    }
}

function gotStream(s) {
    localStream = s;
    debug('Got local stream.');
    shouldBe('localStream.getAudioTracks().length', '1');
    shouldBe('localStream.getVideoTracks().length', '1');

    shouldBeTrue('checkIdAttribute(localStream.id)');
    testConstructor(localStream);
}

function testConstructor(s) {
    stream = s;
    shouldThrow('new MediaStream(document)');
    shouldThrow('new MediaStream([document])');
    shouldThrow('new MediaStream([stream.getAudioTracks()[0], document])');
    shouldThrow('new MediaStream([null])');
    shouldThrow('new MediaStream([undefined])');
    shouldThrow('new MediaStream(null)');
    shouldThrow('new MediaStream(undefined)');

    verifyStream(new MediaStream(), 0, 0);
    verifyStream(new MediaStream([]), 0, 0);

    verifyStream(new MediaStream(s), s.getAudioTracks().length, s.getVideoTracks().length);

    verifyStream(new MediaStream([s.getAudioTracks()[0]]), 1, 0);
    verifyStream(new MediaStream([s.getVideoTracks()[0]]), 0, 1);
    verifyStream(new MediaStream([s.getAudioTracks()[0], s.getVideoTracks()[0]]), 1, 1);
    verifyStream(new MediaStream([s.getVideoTracks()[0], s.getAudioTracks()[0], s.getVideoTracks()[0]]), 1, 1);

    s.oninactive = function () {
        verifyStream(new MediaStream([s.getAudioTracks()[0]]), 0, 0);
        verifyStream(new MediaStream([s.getVideoTracks()[0]]), 0, 0);
        verifyStream(new MediaStream([s.getAudioTracks()[0], s.getVideoTracks()[0]]), 0, 0);
        verifyStream(new MediaStream(s), 0, 0);

        finishJSTest();
    };

    s.getAudioTracks()[0].stop();
    s.getVideoTracks()[0].stop();
}

function verifyStream(s, numAudioTracks, numVideoTracks) {
    newStream = s;
    nAudio = numAudioTracks;
    nVideo = numVideoTracks;

    debug('Stream constructed');
    shouldBeNonNull(newStream);
    shouldBeDefined(newStream);
    shouldBe('newStream.constructor.name', "'MediaStream'");
    shouldBe('newStream.getAudioTracks().length', 'nAudio');
    shouldBe('newStream.getVideoTracks().length', 'nVideo');

    if (!nAudio && !nVideo)
        shouldBe('newStream.active', 'false');
    else
        shouldBe('newStream.active', 'true');

    shouldBeTrue('checkIdAttribute(newStream.id)');
}

testInterfaces();
getUserMedia({video:true, audio:true}, gotStream);

window.jsTestIsAsync = true;
window.successfullyParsed = true;
</script>
</body>
</html>
